restructure send-reset-token folders
