--- src/services/send-password-reset-link/index.js
@@ -1,7 +1,0 @@
-const generatePasswordResetLink = require('./generate-password-reset-link')
-const sendPasswordResetLink = require('./send-password-reset-link')
-
-module.exports = {
-	generatePasswordResetLink,
-	sendPasswordResetLink
-}
