+++ src/services/users/index.js
@@ -10,6 +10,9 @@ const { AppError } = require("../../lib")
 const generateAccessToken = require("./generate-access-token")
 const generateRefreshToken = require("./generate-refresh-token")
 const updateRefreshToken = require("./update-refresh-token")
+const { generatePasswordResetLink 
+	,sendPasswordResetLink 
+} = require('./send-password-reset-link')
 
 const {
 	RefreshTokenIsInvalidError,
@@ -221,5 +224,7 @@ module.exports = {
 	signInUserByEmail,
 	inviteUser,
 	getNewTokens,
-	getInvitedUserDetail
+	getInvitedUserDetail,
+	generatePasswordResetLink,
+	sendPasswordResetLink
 }
